refactor(idl-gen): extract extension discovery into its own module
Splits `spel-framework-core::extension` out of idl_gen.rs (~300 lines
moved). Holds the 5 discovery functions added in 99a26ea plus their
tests. Public API: `discover_extension_instructions` and
`discover_extension_instruction_attrs`; the three internal helpers
tighten to module-private. Feature-gated identically to `idl_gen`.

Updates the three call sites in `spel-framework-macros` to the new
path. No behavior change.
