+++ spel-framework-core/src/lib.rs
@@ -15,6 +15,8 @@ pub mod validation;
 #[cfg(feature = "idl-gen")]
 pub mod account_types;
 #[cfg(feature = "idl-gen")]
+pub mod extension;
+#[cfg(feature = "idl-gen")]
 pub mod idl_gen;
 
 pub mod prelude {
